added Preparatory Work for VHI
