+++ tools/last_updates.csv
@@ -1,3 +1,4 @@
 Product,LastSceneDate,RunDate,Status
 NDVI-MAX,2024-02-15,2024-02-15,complete
 ch.swisstopo.swisseo_s2-sr_v100,2023-08-22,2024-03-21T12:09,RUNNING
+ch.swisstopo.swisseo_vhi_v100,2020-07-18,2020-07-18,RUNNING
